graph and graph_grammar
