+++ app/app.py
@@ -13,7 +13,7 @@ from control_optimisation import (create_grab_criterion_fun, create_traj_fun, ge
 
 from rostok.criterion.flags_simualtions import (FlagMaxTime, FlagNotContact, FlagSlipout)
 from rostok.graph_generators.mcts_helper import (make_mcts_step, prepare_mcts_state_and_helper)
-from rostok.graph_grammar.node import GraphGrammar
+from rostok.graph_grammar.graph_grammar import GraphGrammar
 from rostok.trajectory_optimizer.control_optimizer import (ConfigRewardFunction, ControlOptimizer)
