+++ examples/brute_force/rule_extention.py
@@ -1,24 +1,53 @@
 import matplotlib.pyplot as plt
+import networkx as nx
 import numpy as np
 import pychrono as chrono
 
-from rostok.block_builder.node_render import *
-from rostok.graph_grammar import node_vocabulary, rule_vocabulary
-from rostok.graph_grammar.node import ROOT, BlockWrapper
+from rostok.block_builder_chrono.block_classes import *
+from rostok.block_builder_chrono.easy_body_shapes import Box
+from rostok.graph_grammar import rule_vocabulary
+from rostok.graph import node_vocabulary
+from rostok.graph.node import BlockWrapper
+from rostok.graph.graph import ROOT
+
+
+def plot_graph(graph):
+    plt.figure()
+    nx.draw_networkx(graph,
+                     pos=nx.kamada_kawai_layout(graph, dim=2),
+                     node_size=800,
+                     labels={n: graph.nodes[n]["Node"].label for n in G})
+    #plt.figure()
+    #nx.draw_networkx(graph, pos=nx.kamada_kawai_layout(G, dim=2), node_size=800)
+    plt.show()
 
 
 def init_extension_rules():
-    # %% Bodies for extansions rules
-    width = [0.25, 0.35, 0.5]
-    length_link = [0.4, 0.6, 0.8]
+    width = [0.5, 0.6, 0.65]
+    alpha = 45
+    alpha_left = [0, 30, 60]
+    alpha_right = [180, 150, 120]
+    length_link = [0.3, 0.6, 0.8, 1]
 
-    flat = list(
-        map(lambda x: BlockWrapper(FlatChronoBody, width_x=x, height_y=0.05, depth_z=0.8), width))
+    flat = list(map(lambda x: BlockWrapper(PrimitiveBody, Box(x, 0.05, 0.8)), width))
 
-    link = list(map(lambda x: BlockWrapper(LinkChronoBody, length_y=x), length_link))
+    link = list(map(lambda x: BlockWrapper(PrimitiveBody, Box(0.1, x, 0.3)), length_link))
 
-    u1 = BlockWrapper(MountChronoBody, width_x=0.1, length_y=0.05)
-    u2 = BlockWrapper(MountChronoBody, width_x=0.2, length_y=0.1)
+    u1 = BlockWrapper(PrimitiveBody, Box(0.1, 0.05, 0.3))
+    u2 = BlockWrapper(PrimitiveBody, Box(0.2, 0.1, 0.3))
+
+    # %% Tranform for extansions rules
+    # z_shift = [-0.3, 0, 0.3]
+    # MOVE_TO_RIGHT_SIDE = []
+    # for i in width:
+    #     for j in z_shift:
+    #         for alpha in alpha_right:
+    #             quat_Y_ang_alpha = chrono.Q_from_AngY(np.deg2rad(alpha))
+    #             ROTATE_TO_ALPHA = FrameTransform([i, 0, j],[quat_Y_ang_alpha.e0,quat_Y_ang_alpha.e1,
+    #                                         quat_Y_ang_alpha.e2,quat_Y_ang_alpha.e3])
+    #             MOVE_TO_RIGHT_SIDE.append(ROTATE_TO_ALPHA)
+
+    # transform_to_right_mount_list = list(map(lambda x: BlockWrapper(ChronoTransform, x), MOVE_TO_RIGHT_SIDE))
 
     def rotation(alpha):
         quat_Y_ang_alpha = chrono.Q_from_AngY(np.deg2rad(alpha))
@@ -37,6 +66,10 @@ def init_extension_rules():
     MOVE_TO_LEFT_SIDE_MINUS = map(lambda x: FrameTransform([-x, 0, -0.3], [1, 0, 0, 0]), width)
     MOVE_TO_LEFT_SIDE_MINUS_ANGLE = map(lambda x: FrameTransform([-x, 0, -0.3], rotation(-30)),
                                         width)
+
+    # quat_Y_ang_alpha = chrono.Q_from_AngY(np.deg2rad(alpha))
+    # ROTATE_TO_ALPHA = FrameTransform([0, 0, 0],[quat_Y_ang_alpha.e0,quat_Y_ang_alpha.e1,
+    #                                         quat_Y_ang_alpha.e2,quat_Y_ang_alpha.e3])
 
     transform_to_right_mount = list(
         map(lambda x: BlockWrapper(ChronoTransform, x), MOVE_TO_RIGHT_SIDE))
@@ -59,6 +92,7 @@ def init_extension_rules():
     transform_to_left_mount_minus_angle = list(
         map(lambda x: BlockWrapper(ChronoTransform, x), MOVE_TO_LEFT_SIDE_MINUS_ANGLE))
     # transform_to_alpha_rotate = BlockWrapper(ChronoTransform, ROTATE_TO_ALPHA)
+
     # %%
     type_of_input = ChronoRevolveJoint.InputType.TORQUE
 
@@ -68,7 +102,6 @@ def init_extension_rules():
     # Nodes
     node_vocab = node_vocabulary.NodeVocabulary()
     node_vocab.add_node(ROOT)
-    node_vocab.create_node("J")
     node_vocab.create_node("L")
     node_vocab.create_node("F")
     node_vocab.create_node("M")
@@ -85,6 +118,7 @@ def init_extension_rules():
     node_vocab.create_node("SMLM")
     node_vocab.create_node("SMLMA")
 
+    #O = Node("O")
     node_vocab.create_node(label="J1", is_terminal=True, block_wrapper=revolve1)
     node_vocab.create_node(label="L1", is_terminal=True, block_wrapper=link[0])
     node_vocab.create_node(label="L2", is_terminal=True, block_wrapper=link[1])
@@ -180,26 +214,41 @@ def init_extension_rules():
 
     rule_vocab.create_rule("InitMechanism_2", ["ROOT"], ["F", "SML", "SMR", "EM", "EM"], 0, 0,
                            [(0, 1), (0, 2), (1, 3), (2, 4)])
-    #rule_vocab.create_rule("InitMechanism_2", ["ROOT"], ["F1", "TL1", "TR1","EM","EM"], 0 , 0,[(0,1),(0,2),(1,3),(2,4)])
-    #rule_vocab.create_rule("InitMechanism_3_R", ["ROOT"], ["F", "SML", "SMRP","SMRM","EM","EM","EM"], 0 , 0,[(0,1),(0,2),(0,3),(1,4),(2,5),(3,6)])
-    # rule_vocab.create_rule("InitMechanism_3_R_A", ["ROOT"], ["F", "SML", "SMRPA","SMRMA","EM","EM","EM"], 0 , 0,[(0,1),(0,2),(0,3),(1,4),(2,5),(3,6)])
-    # rule_vocab.create_rule("InitMechanism_3_L", ["ROOT"], ["F", "SMLP","SMLM", "SMR","EM","EM","EM"], 0 , 0, [(0,1),(0,2),(0,3),(1,4),(2,5),(3,6)])
-    # rule_vocab.create_rule("InitMechanism_3_L_A", ["ROOT"], ["F", "SMLPA","SMLMA", "SMR","EM","EM","EM"], 0 , 0, [(0,1),(0,2),(0,3),(1,4),(2,5),(3,6)])
-    # rule_vocab.create_rule("InitMechanism_4", ["ROOT"], ["F", "SMLP","SMLM", "SMRP","SMRM","EM","EM","EM","EM"], 0 , 0, [(0,1),(0,2),(0,3),(0,4),(1,5),(2,6),(3,7),(4,8)])
-    # rule_vocab.create_rule("InitMechanism_4_A", ["ROOT"], ["F", "SMLPA","SMLMA", "SMRPA","SMRMA","EM","EM","EM","EM"], 0 , 0, [(0,1),(0,2),(0,3),(0,4),(1,5),(2,6),(3,7),(4,8)])
+    rule_vocab.create_rule("InitMechanism_3_R", ["ROOT"],
+                           ["F", "SML", "SMRP", "SMRM", "EM", "EM", "EM"], 0, 0, [(0, 1), (0, 2),
+                                                                                  (0, 3), (1, 4),
+                                                                                  (2, 5), (3, 6)])
+    rule_vocab.create_rule("InitMechanism_3_R_A", ["ROOT"],
+                           ["F", "SML", "SMRPA", "SMRMA", "EM", "EM", "EM"], 0, 0, [(0, 1), (0, 2),
+                                                                                    (0, 3), (1, 4),
+                                                                                    (2, 5), (3, 6)])
+    rule_vocab.create_rule("InitMechanism_3_L", ["ROOT"],
+                           ["F", "SMLP", "SMLM", "SMR", "EM", "EM", "EM"], 0, 0, [(0, 1), (0, 2),
+                                                                                  (0, 3), (1, 4),
+                                                                                  (2, 5), (3, 6)])
+    rule_vocab.create_rule("InitMechanism_3_L_A", ["ROOT"],
+                           ["F", "SMLPA", "SMLMA", "SMR", "EM", "EM", "EM"], 0, 0, [(0, 1), (0, 2),
+                                                                                    (0, 3), (1, 4),
+                                                                                    (2, 5), (3, 6)])
+    rule_vocab.create_rule("InitMechanism_4", ["ROOT"],
+                           ["F", "SMLP", "SMLM", "SMRP", "SMRM", "EM", "EM", "EM", "EM"], 0, 0,
+                           [(0, 1), (0, 2), (0, 3), (0, 4), (1, 5), (2, 6), (3, 7), (4, 8)])
+    rule_vocab.create_rule("InitMechanism_4_A", ["ROOT"],
+                           ["F", "SMLPA", "SMLMA", "SMRPA", "SMRMA", "EM", "EM", "EM", "EM"], 0, 0,
+                           [(0, 1), (0, 2), (0, 3), (0, 4), (1, 5), (2, 6), (3, 7), (4, 8)])
     rule_vocab.create_rule("FingerUpper", ["EM"], ["J1", "L", "EM"], 0, 2, [(0, 1), (1, 2)])
 
     rule_vocab.create_rule("TerminalFlat1", ["F"], ["F1"], 0, 0)
     rule_vocab.create_rule("TerminalFlat2", ["F"], ["F2"], 0, 0)
-    #rule_vocab.create_rule("TerminalFlat3", ["F"], ["F3"], 0 , 0)
+    rule_vocab.create_rule("TerminalFlat3", ["F"], ["F3"], 0, 0)
 
     rule_vocab.create_rule("TerminalL1", ["L"], ["L1"], 0, 0)
     rule_vocab.create_rule("TerminalL2", ["L"], ["L2"], 0, 0)
-    #rule_vocab.create_rule("TerminalL3", ["L"], ["L3"], 0 , 0)
+    rule_vocab.create_rule("TerminalL3", ["L"], ["L3"], 0, 0)
 
     rule_vocab.create_rule("TerminalTransformRight1", ["SMR"], ["TR1"], 0, 0)
     rule_vocab.create_rule("TerminalTransformRight2", ["SMR"], ["TR2"], 0, 0)
-    #rule_vocab.create_rule("TerminalTransformRight3", ["SMR"], ["TR3"], 0 , 0)
+    rule_vocab.create_rule("TerminalTransformRight3", ["SMR"], ["TR3"], 0, 0)
 
     rule_vocab.create_rule("TerminalTransformRightPlus1", ["SMRP"], ["TRP1"], 0, 0)
     rule_vocab.create_rule("TerminalTransformRightPlus2", ["SMRP"], ["TRP2"], 0, 0)
@@ -219,7 +268,7 @@ def init_extension_rules():
 
     rule_vocab.create_rule("TerminalTransformLeft1", ["SML"], ["TL1"], 0, 0)
     rule_vocab.create_rule("TerminalTransformLeft2", ["SML"], ["TL2"], 0, 0)
-    #rule_vocab.create_rule("TerminalTransformLeft3", ["SML"], ["TL3"], 0 , 0)
+    rule_vocab.create_rule("TerminalTransformLeft3", ["SML"], ["TL3"], 0, 0)
 
     rule_vocab.create_rule("TerminalTransformLeftPlus1", ["SMLP"], ["TLP1"], 0, 0)
     rule_vocab.create_rule("TerminalTransformLeftPlus2", ["SMLP"], ["TLP2"], 0, 0)
@@ -239,7 +288,6 @@ def init_extension_rules():
 
     rule_vocab.create_rule("TerminalEndLimb1", ["EM"], ["U1"], 0, 0)
     rule_vocab.create_rule("TerminalEndLimb2", ["EM"], ["U2"], 0, 0)
-    #rule_vocab.create_rule("TerminalJoint", ["J"], ["J1"], 0 , 0)
 
     list_J = node_vocab.get_list_of_nodes(["J1"])
     list_RM = node_vocab.get_list_of_nodes([
