+++ examples/reward_grab_mechanism.py
@@ -1,16 +1,15 @@
-import random
-import pickle 
+import pickle
 
 import pychrono as chrono
-from example_vocabulary import (get_terminal_graph_three_finger)
+from example_vocabulary import get_terminal_graph_three_finger
 
 import rostok.virtual_experiment.simulation_step as step
-from rostok.block_builder.envbody_shapes import Box
-from rostok.block_builder.node_render import (ChronoBodyEnv, DefaultChronoMaterial, FrameTransform)
+from rostok.block_builder_chrono.block_classes import (ChronoEasyShapeObject, DefaultChronoMaterial,
+                                                       FrameTransform)
+from rostok.block_builder_chrono.easy_body_shapes import Box
 from rostok.criterion.criterion_calc import criterion_calc
-from rostok.criterion.flags_simualtions import FlagMaxTime, FlagNotContact, FlagSlipout
-from rostok.graph_grammar.node import BlockWrapper
-from rostok.graph_grammar.nodes_division import nodes_division, sort_left_right
+from rostok.criterion.flags_simualtions import (FlagMaxTime, FlagNotContact, FlagSlipout)
+from rostok.graph.node import BlockWrapper
 from rostok.trajectory_optimizer.control_optimizer import num_joints
 from rostok.trajectory_optimizer.trajectory_generator import \
     create_torque_traj_from_x
@@ -28,12 +27,11 @@ number_trq = num_joints(graph)
 const_torque_koef = [0, 0, 0, 0, -1, 6]
 arr_trj = create_torque_traj_from_x(graph, const_torque_koef, MAX_TIME, TIME_STEP)
 
-
 # Create object to grasp with material props
 mat = DefaultChronoMaterial()
 mat.Friction = 0.65
 mat.DampingF = 0.65
-obj = BlockWrapper(ChronoBodyEnv,
+obj = BlockWrapper(ChronoEasyShapeObject,
                    shape=Box(),
                    material=mat,
                    pos=FrameTransform([0, 0.4, 0], [0, -0.048, 0.706, 0.706]))
@@ -57,8 +55,6 @@ f.close()
 # Weight coefficients for reward function
 WEIGHTS = [5, 10, 2]
 
-
 # Calculate reward value
 reward = criterion_calc(sim_output, WEIGHTS)
 print(reward)
-
