+++ rostok/graph_grammar/rule_vocabulary.py
@@ -4,8 +4,9 @@ import matplotlib.pyplot as plt
 import networkx as nx
 import numpy as np
 
-from rostok.graph_grammar.node import ROOT, GraphGrammar, Rule
-from rostok.graph_grammar.node_vocabulary import NodeVocabulary
+from rostok.graph.graph import ROOT
+from rostok.graph.node_vocabulary import NodeVocabulary
+from rostok.graph_grammar.graph_grammar import GraphGrammar, Rule
 
 
 class RuleVocabulary():
