+++ rostok/trajectory_optimizer/control_optimizer.py
@@ -5,7 +5,7 @@ import pychrono as chrono
 from scipy.optimize import direct, dual_annealing, shgo
 
 from rostok.block_builder_chrono.block_classes import NodeFeatures
-from rostok.graph_grammar.node import GraphGrammar
+from rostok.graph_grammar.graph_grammar import GraphGrammar
 from rostok.virtual_experiment.robot import Robot
 from rostok.virtual_experiment.simulation_step import (SimOut, SimulationStepOptimization)
