+++ rostok/virtual_experiment/simulation_step.py
@@ -8,7 +8,8 @@ from rostok.block_builder_chrono.block_classes import (BuildingBody, ChronoRevol
 from rostok.block_builder_chrono.blocks_utils import (FrameTransform, OriginWorldFrame)
 from rostok.block_builder_chrono.chrono_system import register_chrono_system
 from rostok.criterion.flags_simualtions import (ConditionStopSimulation, FlagStopSimualtions)
-from rostok.graph_grammar.node import BlockWrapper, GraphGrammar
+from rostok.graph.node import BlockWrapper
+from rostok.graph_grammar.graph_grammar import GraphGrammar
 from rostok.virtual_experiment.auxilarity_sensors import RobotSensor
 from rostok.virtual_experiment.robot import Robot
